chore: define run_cargo_check
